Biproportional apportionment

Method to proportionally allocate seatsamong parties (or lists) and districts (or entities, regions).

biproportional(
  votes_matrix,
  district_seats,
  quorum_districts = 0,
  quorum_total = 0,
  use_list_votes = TRUE,
  method = "round"
)

Arguments

votes_matrix
Vote count matrix with votes by party in rows and votes by district in columns

district_seats
Vector defining the number of seats per district. Must be the same length as ncol(votes_matrix). If the number of seats per district should be assigned according to the number of votes (not the general use case), a single number for the total number of seats can be used.

quorum_districts
Vote threshold a party must reach in at least one district. Used as quota of total votes within a district if less than 1 otherwise as number of votes.

quorum_total
Vote threshold a party must reach for all votes cast. Used as quota of total votes if less than 1 otherwise as number of votes.

use_list_votes
By default (TRUE) it's assumed that each voter in a district has as many votes as there are seats in a district. Set to FALSE if votes_matrix shows the number of voters.

method
Defines how seats in upper and lower apportionment are assigned. The default "round" for the Sainte-Lague/Webster method is the standard for biproportional apportionment. See proporz for a list of available methods. For a different method for upper and lower apportionment use a vector with two entries. It is also possible to provide a function that works like base::round(x) (i.e. can handle a matrix).

Details

Each party nominates a candidate list for every district. The voters vote for the parties of their district. The seat allocation is calculated in two steps:

In the so called upper apportionment the number of seats for each party (over all districts) is determined. Normally, the number of seats for each region are defined before the election and are independent of the vote counts.

In the so called lower apportionment the seats are distributed to the regional party list respecting the results from the upper apportionment.

Parties failing to reach at least one quorum cannot get seats.

Note

The iterative process in the lower apportionment is only guaranteed to terminate with Sainte-Laguë/Webster method.

See also

Examples

votes_df = unique(zug2018[c("list_id", "entity_id", "list_votes")])
votes_matrix = pivot_to_matrix(votes_df)
votes_matrix

#>        entity_id
#> list_id  1701 1702 1703 1704 1705 1706 1707 1708 1709 1710  1711
#>       1  2993    0    0    0    0    0    0    0    0    0     0
#>       2  8108 4687 1584  531  279  477 2363 3860 1481   91 22023
#>       3 19389 9334 4807 1946  396 2844 3523 4702 3310  812 21343
#>       4 14814 6691 4005  826  379 1654 2842 2624 2713  461 33789
#>       5  4486 2270  621  198    0  361  728  465  925    0 10131
#>       6 15695 4705 1750   84    0   51  627 1106 1563  302 21794
#>       7 21298 8178 2875 1336  399 1450 3715 2610 4063  344 26798

#>        entity_id
#> list_id  1701 1702 1703 1704 1705 1706 1707 1708 1709 1710  1711
#>       1  2993    0    0    0    0    0    0    0    0    0     0
#>       2  8108 4687 1584  531  279  477 2363 3860 1481   91 22023
#>       3 19389 9334 4807 1946  396 2844 3523 4702 3310  812 21343
#>       4 14814 6691 4005  826  379 1654 2842 2624 2713  461 33789
#>       5  4486 2270  621  198    0  361  728  465  925    0 10131
#>       6 15695 4705 1750   84    0   51  627 1106 1563  302 21794
#>       7 21298 8178 2875 1336  399 1450 3715 2610 4063  344 26798

distr_df = unique(zug2018[c("entity_id", "election_mandates")])
district_seats = setNames(distr_df$election_mandates, distr_df$entity_id)
district_seats

#> 1701 1702 1703 1704 1705 1706 1707 1708 1709 1710 1711 
#>   15   10    6    3    2    4    7    6    6    2   19 

#> 1701 1702 1703 1704 1705 1706 1707 1708 1709 1710 1711
#>   15   10    6    3    2    4    7    6    6    2   19

biproportional(votes_matrix, district_seats, 0.05, 0.03)

#>        entity_id
#> list_id 1701 1702 1703 1704 1705 1706 1707 1708 1709 1710 1711
#>       1    0    0    0    0    0    0    0    0    0    0    0
#>       2    2    1    1    0    0    0    1    2    1    0    3
#>       3    3    3    2    1    1    2    2    2    1    1    3
#>       4    2    2    1    1    0    1    2    1    1    1    5
#>       5    1    1    0    0    0    0    0    0    0    0    2
#>       6    3    1    1    0    0    0    0    0    1    0    3
#>       7    4    2    1    1    1    1    2    1    2    0    3
#> attr(,"divisors")
#> attr(,"divisors")$districts
#>      1701      1702      1703      1704      1705      1706      1707      1708 
#> 5745.5376 3586.0000 2607.0000 1435.0000  743.0312 1709.0000 1816.8750 2561.0000 
#>      1709      1710      1711 
#> 2342.0000  725.5000 7275.1719 
#> 
#> attr(,"divisors")$parties
#>         1         2         3         4         5         6         7 
#> 1.0000000 0.8828125 1.0000000 1.0316329 0.8750000 1.0000000 1.0537109 
#> 
#> attr(,"class")
#> [1] "proporz_matrix" "matrix"         "array"         

#>         entity_id
#> list_id 1701 1702 1703 1704 1705 1706 1707 1708 1709 1710 1711
#>       1    0    0    0    0    0    0    0    0    0    0    0
#>       2    2    1    1    0    0    0    1    2    1    0    3
#>       3    3    3    2    1    1    2    2    2    1    1    3
#>       4    2    2    1    1    0    1    2    1    1    1    5
#>       5    1    1    0    0    0    0    0    0    0    0    2
#>       6    3    1    1    0    0    0    0    0    1    0    3
#>       7    4    2    1    1    1    1    2    1    2    0    3

# Different method for upper apportionment
# and using number of voters instead of list votes
f19_matrix = pivot_to_matrix(finland2019$votes_df)
f19_distr_seats = setNames(
    finland2019$district_seats_df$election_mandates,
    finland2019$district_seats_df$entity_id)

biproportional(f19_matrix, f19_distr_seats,
               use_list_votes = FALSE,
               method = c("floor", "round"))

#>         entity_id
#> list_id  HAEM HEL KAA KES LAP OUL PIR SAT SKA UUS VAA VAR
#>   Asyl      0   0   0   0   0   0   0   0   0   0   0   0
#>   E117      0   0   0   0   0   0   0   0   0   0   0   0
#>   E118      0   0   0   0   0   0   0   0   0   0   0   0
#>   E154      0   0   0   0   0   0   0   0   0   0   0   0
#>   E168      0   0   0   0   0   0   0   0   0   0   0   0
#>   E169      0   0   0   0   0   0   0   0   0   0   0   0
#>   E185      0   0   0   0   0   0   0   0   0   0   0   0
#>   E191      0   0   0   0   0   0   0   0   0   0   0   0
#>   E192      0   0   0   0   0   0   0   0   0   0   0   0
#>   E287      0   0   0   0   0   0   0   0   0   0   0   0
#>   E491      0   0   0   0   0   0   0   0   0   0   0   0
#>   E492      0   0   0   0   0   0   0   0   0   0   0   0
#>   E493      0   0   0   0   0   0   0   0   0   0   0   0
#>   EOP       0   0   0   0   0   0   0   0   0   0   0   0
#>   FP        0   0   0   0   0   0   0   0   0   0   0   0
#>   IP        0   0   0   0   0   0   0   0   0   0   0   0
#>   KD        1   0   1   1   0   1   1   0   1   1   1   0
#>   KESK      1   1   3   2   2   6   2   1   3   2   3   2
#>   KOK       3   5   3   1   1   2   4   1   2   7   2   4
#>   KP        0   0   0   0   0   0   0   0   0   0   0   0
#>   KTP       0   0   0   0   0   0   0   0   0   0   0   0
#>   LIBE      0   0   0   0   0   0   0   0   0   0   0   0
#>   LII       0   0   0   0   0   0   0   0   0   0   0   0
#>   LIIK      0   0   0   0   0   0   0   0   0   0   0   0
#>   LN        0   0   0   0   0   0   0   0   0   0   0   0
#>   LNLY      0   0   0   0   0   0   0   0   0   0   0   0
#>   LNY       0   0   0   0   0   0   0   0   0   0   0   0
#>   LNYL      0   0   0   0   0   0   0   0   0   0   0   0
#>   Nyt       0   0   0   0   0   0   0   0   0   1   0   0
#>   NYT       0   1   0   0   0   0   0   0   0   0   0   0
#>   PEL       0   0   0   0   0   0   0   0   0   0   0   0
#>   PIR       0   1   0   0   0   0   0   0   0   0   0   0
#>   PS        3   3   3   2   1   4   3   2   3   6   3   3
#>   REF       0   0   0   0   0   0   0   0   0   0   0   0
#>   Reform    0   0   0   0   0   0   0   0   0   0   0   0
#>   RKP       0   1   0   0   0   0   0   0   0   4   3   1
#>   RLI       0   0   0   0   0   0   0   0   0   0   0   0
#>   SDP       4   3   4   2   1   2   4   2   3   7   2   3
#>   SIN       0   0   0   0   0   0   0   0   1   1   0   0
#>   SKE       0   0   0   0   0   0   0   0   0   0   0   0
#>   SKP       0   0   0   0   0   0   0   0   0   0   0   0
#>   STL       0   0   0   0   0   0   0   0   0   0   0   0
#>   VAS       1   2   1   1   1   2   2   1   1   2   1   2
#>   VIHR      1   5   2   1   1   1   3   1   1   5   1   2
#> attr(,"divisors")
#> attr(,"divisors")$districts
#>     HAEM      HEL      KAA      KES      LAP      OUL      PIR      SAT 
#> 14001.88 18399.13 14410.00 15438.00 14258.00 14105.17 15126.41 14019.00 
#>      SKA      UUS      VAA      VAR 
#> 14596.33 15343.18 16505.43 14892.57 
#> 
#> attr(,"divisors")$parties
#>      Asyl      E117      E118      E154      E168      E169      E185      E191 
#> 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 
#>      E192      E287      E491      E492      E493       EOP        FP        IP 
#> 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 
#>        KD      KESK       KOK        KP       KTP      LIBE       LII      LIIK 
#> 0.8117065 1.0078125 0.9946907 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 
#>        LN      LNLY       LNY      LNYL       Nyt       NYT       PEL       PIR 
#> 1.0000000 1.0000000 1.0000000 1.0000000 1.2500000 1.0000000 1.0000000 0.6250000 
#>        PS       REF    Reform       RKP       RLI       SDP       SIN       SKE 
#> 1.0156250 1.0000000 1.0000000 0.9163910 1.0000000 0.9725345 0.5625000 1.0000000 
#>       SKP       STL       VAS      VIHR 
#> 1.0000000 1.0000000 1.0000000 0.9773330 
#> 
#> attr(,"class")
#> [1] "proporz_matrix" "matrix"         "array"