Create weighted votes matrix

Weight list votes by dividing the votes matrix entries by the number of seats per district. This method is used in upper_apportionment() if use_list_votes is TRUE (default).

Usage

weight_list_votes(votes_matrix, district_seats)

Arguments

votes_matrix

votes matrix

district_seats

seats per district, vector with same length as ncol(votes_matrix)

Value

the weighted votes_matrix

Note

The weighted votes are not rounded. Matrix and vector names are ignored.

Examples

weight_list_votes(uri2020$votes_matrix, uri2020$seats_vector)
#>       Altdorf  Bürglen Erstfeld Schattdorf
#> CVP  764.7333 403.1429 384.8333   532.6667
#> SPGB 793.8667 229.4286 284.1667   288.8889
#> FDP  614.2000 223.8571 157.6667   329.0000
#> SVP  517.0667 420.7143 262.1667   388.6667