Calculate lower apportionment

Iterate and change column and row divisors such that the row and column sums of the seats matrix satisfies the constraints given by the upper apportionment.

lower_apportionment(votes_matrix, seats_cols, seats_rows, method = "round")

Arguments

votes_matrix

votes matrix

seats_cols

number of seats per column (districts/regions), predetermined or calculated with upper_apportionment().

seats_rows

number of seats per row (parties/lists), calculated with upper_apportionment().

method

Apportion method that defines how seats are assigned. The default "round" for the Sainte-Laguë/Webster method is the standard for biproportional apportionment and the only method guaranteed to terminate. See proporz() for other methods. It is also possible to provide a function that round a vector or matrix.

Value

A seat matrix with district (columns) and party (rows) divisors stored in attributes.

Details

The result is obtained by an iterative process ('Alternate Scaling Algorithm', see Reference). Initially, for each district a divisor is chosen using the highest averages method for the votes allocated to each regional party list in this region. For each party a party divisor is initialized with 1.

Effectively, the objective of the iterative process is to modify the regional divisors and party divisors so that the number of seats in each regional party list equals the number of their votes divided by both the regional and the party divisors.

The following two correction steps are executed until this objective is satisfied:

modify the party divisors such that the apportionment within each party is correct with the chosen rounding method,

modify the regional divisors such that the apportionment within the region is correct with the chosen rounding method.

References

Oelbermann, K. F. (2016). Alternate scaling algorithm for biproportional divisor methods. Mathematical Social Sciences, 80, 25-32.

See also

Examples

votes_matrix = matrix(c(123,912,312,45,714,255,815,414,215), nrow = 3)
district_seats = c(7,5,8)
party_seats = c(5,11,4)

lower_apportionment(votes_matrix, district_seats, party_seats)
#>      [,1] [,2] [,3]
#> [1,]    1    0    4
#> [2,]    4    4    3
#> [3,]    2    1    1
#> attr(,"divisors")
#> attr(,"divisors")$districts
#> [1] 207.8115 180.2500 161.5000
#> 
#> attr(,"divisors")$parties
#> [1] 1.156250 1.007812 1.000000
#>