Biproportional apportionment with data frames

Method to proportionally allocate seatsamong parties/lists and districts/regions/entities ("Doppelter Pukelsheim").

pukelsheim(
  votes_df,
  district_seats_df,
  new_seats_col = "seats",
  quorum_districts = 0,
  quorum_total = 0,
  use_list_votes = TRUE
)

Arguments

votes_df

data.frame (long format) with 3 columns (actual colnames can differ):

party id/name

district id/name

votes

district_seats_df

data.frame with 2 columns (actual colnames can differ):

district id/name

number of seats for a district

new_seats_col

name of the new column

quorum_districts

Vote threshold a party must reach in at least one district. Used as quota of total votes within a district if less than 1 otherwise as number of votes.

quorum_total

Vote threshold a party must reach for all votes cast. Used as quota of total votes if less than 1, otherwise as number of votes.

use_list_votes

By default (TRUE) it's assumed that each voter in a district has as many votes as there are seats in a district. Set to FALSE if votes_df shows the number of voters.

Details

Each party nominates a candidate list for every district. The voters vote for the parties of their district. The seat allocation is calculated in two steps:

In the so called upper apportionment the number of seats for each party (over all districts) is determined.

In the so called lower apportionment the seats are distributed to the regional party list respecting the results from the upper apportionment.

Parties failing to reach at least one quorum have their votes set to zero.

If you want to other apportion methods than Sainte-Laguë use biproportional.

See also

Examples

# Zug 2018
votes_df = unique(zug2018[c("list_id", "entity_id", "list_votes")])
district_seats_df = unique(zug2018[c("entity_id", "election_mandates")])

seats_df = pukelsheim(votes_df,
                      district_seats_df,
                      quorum_districts = 0.05,
                      quorum_total = 0.03)

head(seats_df)
#>   list_id entity_id list_votes seats
#> 1       2      1701       8108     2
#> 2       1      1701       2993     0
#> 3       3      1701      19389     3
#> 4       4      1701      14814     2
#> 5       5      1701       4486     1
#> 6       6      1701      15695     3
#>   list_id entity_id list_votes seats
#> 1       2      1701       8108     2
#> 2       1      1701       2993     0
#> 3       3      1701      19389     3
#> 4       4      1701      14814     2
#> 5       5      1701       4486     1
#> 6       6      1701      15695     3

# Finland 2019
finland19_result = pukelsheim(finland2019$votes_df,
                             finland2019$district_seats_df,
                             new_seats_col = "mandates",
                             use_list_votes = FALSE)
tail(finland19_result[order(finland19_result$mandates),])
#>     list_id entity_id list_votes mandates
#> 21     VIHR       HEL      90662        5
#> 411      PS       UUS      86691        5
#> 420    VIHR       UUS      73626        5
#> 226    KESK       OUL      78486        6
#> 414     SDP       UUS      97107        6
#> 405     KOK       UUS     114243        7
#> list_id entity_id list_votes mandates
#>    VIHR       HEL      90662        5
#>      PS       UUS      86691        5
#>    VIHR       UUS      73626        5
#>    KESK       OUL      78486        6
#>     SDP       UUS      97107        6
#>     KOK       UUS     114243        7