Election Data for the Cantonal Council of Uri (2020)

Example election data from the 2020 election for the cantonal council of Uri (Landrat) in Switzerland. The data has been extracted from the report "Landratswahlen 2020: Statistische Auswertung".

Usage

uri2020

Format

List containing:

votes_matrix the number of votes for each party and district (4 rows, 4 columns)

seats_vector with the number of seats per district (length 4)

Source

https://www.ur.ch/abstimmungen/termine/9322